Unit tests überarbeitung
Unit tests werden neu überarbeitet für repositories

+++ src/main/java/at/fhtw/swen3/persistence/repositories/ParcelRepository.java
@@ -5,8 +5,8 @@ import at.fhtw.swen3.persistence.entity.RecipientEntity;
 import io.swagger.models.auth.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParcelRepository extends JpaRepository<ParcelEntity, Integer> {
-    ParcelEntity findById(int id);
+public interface ParcelRepository extends JpaRepository<ParcelEntity, Long> {
+    ParcelEntity findById(long id);
     ParcelEntity findByTrackingId(String trackingId);
     ParcelEntity findBySender(RecipientEntity sender);
     ParcelEntity findByRecipient(RecipientEntity recipient);
